feat: updated build script

+++ build.sh
@@ -1,7 +1,14 @@
-#/usr/bin/env bash
+#!/usr/bin/env bash
+
+typ=""
+target="$2"
+if [ $1 = "go" ] || [ $1 = "py" ]; then
+    typ="$1"
+else
+    target="$1"
+fi
 
 spider_dir="./bin"
-src_dir="./src"
 target_dir="./releases"
 
 build_time=`date +%Y%m%d.%H:%M:%S`
@@ -13,6 +20,7 @@ go_os="linux"
 go_clipkg="github.com/voiladev/VoilaCrawler/pkg/cli"
 go_ldflags="-X $go_clipkg.buildTime=$build_time -X $go_clipkg.buildCommit=$build_commit -X $go_clipkg.buildBranch=$build_branch"
 
+py_src_dir="./src"
 py_loglevel="ERROR"
 
 # buildGo generate executable binary
@@ -31,39 +39,51 @@ buildPy() {
     path=$1
     name=$2
     echo "PY BUILD $path => $target_dir/$name"
-    pyinstaller --onefile --paths $src_dir --distpath=$target_dir --workpath=/tmp/pycrawler --specpath=/tmp/pycrawler -n $name --log-level=$py_loglevel $path
+    pyinstaller --onefile --paths $py_src_dir --distpath=$target_dir --workpath=/tmp/pycrawler --specpath=/tmp/pycrawler -n $name --log-level=$py_loglevel $path
 }
 
-rm -rf $target_dir 
-
-for domain in `ls $spider_dir`; do
-    root="$spider_dir/$domain"
-    if [ ! -d $root ]; then
-        continue
+build() {
+    path=$1
+    name=$2
+    if [ "$typ" = "go" ] || [ "$typ" = "" ]; then
+        if [ -f "$path/main.go" ]; then
+            buildGo "$path" "$name"
+        fi
     fi
-    for sub in `ls $root`; do
-        if [ ! -d "$root/$sub" ]; then
+
+    if [ "$typ" = "py" ] || [ "$typ" = "" ]; then
+        if [ -f "$path/main.py" ]; then
+            buildPy "$path/main.py" "$name.py"
+        fi
+    fi
+}
+
+if [ -z "$target" ];  then
+    rm -rf $target_dir 
+    
+    for domain in `ls $spider_dir`; do
+        root="$spider_dir/$domain"
+        if [ ! -d $root ]; then
             continue
         fi
-
-        if [ -f "$root/$sub/main.go" ]; then
-            buildGo "$root/$sub" "$domain.$sub.bin"
-        fi
-        if [ -f "$root/$sub/main.py" ]; then
-            buildPy "$root/$sub/main.py" "$domain.$sub.bin.py"
-        fi
-
-        for sub2 in `ls "$root/$sub"`; do
-            if [ ! -d "$root/$sub/$sub2" ]; then
+        for sub in `ls $root`; do
+            if [ ! -d "$root/$sub" ]; then
                 continue
             fi
 
-            if [ -f "$root/$sub/$sub2/main.go" ]; then
-                buildGo "$root/$sub/$sub2" "$domain.$sub.$sub2.bin"
-            fi
-            if [ -f "$root/$sub/$sub2/main.py" ]; then
-                buildPy "$root/$sub/$sub2" "$domain.$sub.$sub2.bin.py"
-            fi
+            build "$root/$sub" "$domain.$sub.bin"
+            for sub2 in `ls "$root/$sub"`; do
+                if [ ! -d "$root/$sub/$sub2" ]; then
+                    continue
+                fi
+
+                build "$root/$sub/$sub2" "$domain.$sub.$sub2.bin"
+            done
         done
     done
-done
+else
+    if [ -d $spider_dir/$target ]; then
+        name="${target//\//.}"
+        build "$spider_dir/$target" "$name.bin"
+    fi
+fi
